Update data with vignettes if the vignette is seen while they're queued
We were doing this, but only for active popups. That meant mobs seen by
one method when there was a full stack up or in-combat wouldn't be
updated with a vignette if it was seen slightly later. This was
particularly an issue with chat-triggered mobs, because the sequence
was generally chat-then-vignette.