Add option to only consider loot the current character can receive
This applies to absolutely everything. E.g. mobs won't get mount
behavior if the mount is for a covenant you don't belong to.

This relies on the API to say whether it'll drop for any spec for the
current character. It also counts existing restrictions I've added to
loot, which will cover covenants.

Refs #162