Workaround for the 10.0 model-interior-alpha bug in the popup models
The character models inside a Model frame don't inherit their parent's
alpha in retail currently. This causes problems with in-combat hiding
of the popups, because they're secure-frames and can't be hidden
directly so I'm just setting their alpha to 0. Work around this for now
by setting the interior model alpha when the animation finishes --
which isn't ideal as there's a perceptible slight jump. If this bug
persists, I might want to also apply an animation to the model itself.

Fixes #188