New scanning method: try to TargetUnit and watch whether it's blocked
It's lower-range than cache-scanning used to be, because you do have to
be within actual targeting range. It works best if the interval is
turned down to 0, but I worry about side-effects so I'm leaving it
configurable with a higher default.

Disabled by default, because this seems risky and incredibly taint-prone.