Fix TomTom waypoints pushing onto the arrow and not clearing after close
Was a logic error: I forgot to remove some "clear the last tomtom
waypoint" code, and so when you weren't replacing waypoints it'd clear
the previous one anyway... and then not put a new one on the crazy
arrow. Then TomTom's default behavior is to set the next-closest
waypoint when one is removed from the crazy arrow, so it looked like it
was just ignoring the replacement setting.

Clearing after close was just that code hadn't been adjusted for
multiple TomTom waypoints, so only the most-recent one could be
auto-cleared.