Show ignored mobs in the LDB dropdown
There was a conflation between user-ignored and system-ignored mobs that
made this behavior confusing. System-ignored mobs are inaccessible to
the player generally, and so shouldn't be shown. However, user-ignored
mobs are probably still relevant and checking "ignore" on a source in
settings will no longer make them all vanish from the lists, though
"enable" will still do that.

This also adds some UI to show that they're ignored -- an X in the list
and a line in the tooltip.